build(nx): convert the lint+build targets from 'nx:run-script' to 'nx:run-commands'
The nx run-script executor only looks at the `package.json` file that's nearest to the relevant `project.json` which defines the target. Since libs/api-schemas has a package.json for npm workspace imports, but NOT the 'scripts', this caused run-script to always fail. Converting the targets to use the 'nx:run-commands' executor allows specifying exact paths and fixes the issue.