build(eslint): add `no-restricted-imports` rule for "_circularDependencyFix"
There was a circular dependency issue flagged by @nx/eslint-plugin which was fixed by copying some types into a file called "_circularDependencyFix" in the `stripe-client` lib. This will be the canonical way of ameliorating such issues in the future (to be fixed by later refactors, of course). To ensure these "_circularDependencyFix" files' exports are _**never**_ imported from outside the project in which the file resides (which would defeat the purpose), an ESLint rule has been added banning cross-project importation of these files.